Pluggable index and aliases wildcard resolver
This commit refactors the code within the IndexNameExpressionResolver
so that a single plugin may register an expression resolver that will
replace the WildcardExpressionResolver.
