+++ server/src/main/java/org/elasticsearch/cluster/metadata/IndexNameExpressionResolver.java
@@ -19,7 +19,6 @@
 
 package org.elasticsearch.cluster.metadata;
 
-import org.elasticsearch.ElasticsearchParseException;
 import org.elasticsearch.action.IndicesRequest;
 import org.elasticsearch.action.support.IndicesOptions;
 import org.elasticsearch.cluster.ClusterState;
@@ -28,9 +27,6 @@ import org.elasticsearch.common.Strings;
 import org.elasticsearch.common.collect.ImmutableOpenMap;
 import org.elasticsearch.common.collect.Tuple;
 import org.elasticsearch.common.regex.Regex;
-import org.elasticsearch.common.time.DateFormatter;
-import org.elasticsearch.common.time.DateMathParser;
-import org.elasticsearch.common.time.DateUtils;
 import org.elasticsearch.common.util.set.Sets;
 import org.elasticsearch.index.Index;
 import org.elasticsearch.index.IndexNotFoundException;
@@ -38,9 +34,6 @@ import org.elasticsearch.index.IndexSettings;
 import org.elasticsearch.indices.IndexClosedException;
 import org.elasticsearch.indices.InvalidIndexNameException;
 
-import java.time.Instant;
-import java.time.ZoneId;
-import java.time.ZoneOffset;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
@@ -59,8 +52,12 @@ import java.util.stream.StreamSupport;
 
 public class IndexNameExpressionResolver {
 
-    private final DateMathExpressionResolver dateMathExpressionResolver = new DateMathExpressionResolver();
-    private final List<ExpressionResolver> expressionResolvers = List.of(dateMathExpressionResolver, new WildcardExpressionResolver());
+    private final List<ExpressionResolver> expressionResolvers;
+
+    public IndexNameExpressionResolver(@Nullable ExpressionResolver wildcardExpressionResolver) {
+        this.expressionResolvers = List.of(DateMathExpressionResolver.INSTANCE,
+            wildcardExpressionResolver == null ? new WildcardExpressionResolver() : wildcardExpressionResolver);
+    }
 
     /**
      * Same as {@link #concreteIndexNames(ClusterState, IndicesOptions, String...)}, but the index expressions and options
@@ -327,7 +324,7 @@ public class IndexNameExpressionResolver {
      */
     public boolean hasIndexOrAlias(String aliasOrIndex, ClusterState state) {
         Context context = new Context(state, IndicesOptions.lenientExpandOpen());
-        String resolvedAliasOrIndex = dateMathExpressionResolver.resolveExpression(aliasOrIndex, context);
+        String resolvedAliasOrIndex = DateMathExpressionResolver.INSTANCE.resolveExpression(aliasOrIndex, context);
         return state.metaData().getAliasAndIndexLookup().containsKey(resolvedAliasOrIndex);
     }
 
@@ -337,7 +334,7 @@ public class IndexNameExpressionResolver {
     public String resolveDateMathExpression(String dateExpression) {
         // The data math expression resolver doesn't rely on cluster state or indices options, because
         // it just resolves the date math to an actual date.
-        return dateMathExpressionResolver.resolveExpression(dateExpression, new Context(null, null));
+        return DateMathExpressionResolver.INSTANCE.resolveExpression(dateExpression, new Context(null, null));
     }
 
     /**
@@ -649,18 +646,6 @@ public class IndexNameExpressionResolver {
         }
     }
 
-    private interface ExpressionResolver {
-
-        /**
-         * Resolves the list of expressions into other expressions if possible (possible concrete indices and aliases, but
-         * that isn't required). The provided implementations can also be left untouched.
-         *
-         * @return a new list with expressions based on the provided expressions
-         */
-        List<String> resolve(Context context, List<String> expressions);
-
-    }
-
     /**
      * Resolves alias/index name expressions with wildcards into the corresponding concrete indices/aliases
      */
@@ -882,157 +867,6 @@ public class IndexNameExpressionResolver {
             } else {
                 return Collections.emptyList();
             }
-        }
-    }
-
-    public static final class DateMathExpressionResolver implements ExpressionResolver {
-
-        private static final DateFormatter DEFAULT_DATE_FORMATTER = DateFormatter.forPattern("uuuu.MM.dd");
-        private static final String EXPRESSION_LEFT_BOUND = "<";
-        private static final String EXPRESSION_RIGHT_BOUND = ">";
-        private static final char LEFT_BOUND = '{';
-        private static final char RIGHT_BOUND = '}';
-        private static final char ESCAPE_CHAR = '\\';
-        private static final char TIME_ZONE_BOUND = '|';
-
-        @Override
-        public List<String> resolve(final Context context, List<String> expressions) {
-            List<String> result = new ArrayList<>(expressions.size());
-            for (String expression : expressions) {
-                result.add(resolveExpression(expression, context));
-            }
-            return result;
-        }
-
-        @SuppressWarnings("fallthrough")
-        String resolveExpression(String expression, final Context context) {
-            if (expression.startsWith(EXPRESSION_LEFT_BOUND) == false || expression.endsWith(EXPRESSION_RIGHT_BOUND) == false) {
-                return expression;
-            }
-
-            boolean escape = false;
-            boolean inDateFormat = false;
-            boolean inPlaceHolder = false;
-            final StringBuilder beforePlaceHolderSb = new StringBuilder();
-            StringBuilder inPlaceHolderSb = new StringBuilder();
-            final char[] text = expression.toCharArray();
-            final int from = 1;
-            final int length = text.length - 1;
-            for (int i = from; i < length; i++) {
-                boolean escapedChar = escape;
-                if (escape) {
-                    escape = false;
-                }
-
-                char c = text[i];
-                if (c == ESCAPE_CHAR) {
-                    if (escapedChar) {
-                        beforePlaceHolderSb.append(c);
-                        escape = false;
-                    } else {
-                        escape = true;
-                    }
-                    continue;
-                }
-                if (inPlaceHolder) {
-                    switch (c) {
-                        case LEFT_BOUND:
-                            if (inDateFormat && escapedChar) {
-                                inPlaceHolderSb.append(c);
-                            } else if (!inDateFormat) {
-                                inDateFormat = true;
-                                inPlaceHolderSb.append(c);
-                            } else {
-                                throw new ElasticsearchParseException("invalid dynamic name expression [{}]." +
-                                    " invalid character in placeholder at position [{}]", new String(text, from, length), i);
-                            }
-                            break;
-
-                        case RIGHT_BOUND:
-                            if (inDateFormat && escapedChar) {
-                                inPlaceHolderSb.append(c);
-                            } else if (inDateFormat) {
-                                inDateFormat = false;
-                                inPlaceHolderSb.append(c);
-                            } else {
-                                String inPlaceHolderString = inPlaceHolderSb.toString();
-                                int dateTimeFormatLeftBoundIndex = inPlaceHolderString.indexOf(LEFT_BOUND);
-                                String mathExpression;
-                                String dateFormatterPattern;
-                                DateFormatter dateFormatter;
-                                final ZoneId timeZone;
-                                if (dateTimeFormatLeftBoundIndex < 0) {
-                                    mathExpression = inPlaceHolderString;
-                                    dateFormatter = DEFAULT_DATE_FORMATTER;
-                                    timeZone = ZoneOffset.UTC;
-                                } else {
-                                    if (inPlaceHolderString.lastIndexOf(RIGHT_BOUND) != inPlaceHolderString.length() - 1) {
-                                        throw new ElasticsearchParseException("invalid dynamic name expression [{}]. missing closing `}`" +
-                                            " for date math format", inPlaceHolderString);
-                                    }
-                                    if (dateTimeFormatLeftBoundIndex == inPlaceHolderString.length() - 2) {
-                                        throw new ElasticsearchParseException("invalid dynamic name expression [{}]. missing date format",
-                                            inPlaceHolderString);
-                                    }
-                                    mathExpression = inPlaceHolderString.substring(0, dateTimeFormatLeftBoundIndex);
-                                    String patternAndTZid =
-                                        inPlaceHolderString.substring(dateTimeFormatLeftBoundIndex + 1, inPlaceHolderString.length() - 1);
-                                    int formatPatternTimeZoneSeparatorIndex = patternAndTZid.indexOf(TIME_ZONE_BOUND);
-                                    if (formatPatternTimeZoneSeparatorIndex != -1) {
-                                        dateFormatterPattern = patternAndTZid.substring(0, formatPatternTimeZoneSeparatorIndex);
-                                        timeZone = DateUtils.of(patternAndTZid.substring(formatPatternTimeZoneSeparatorIndex + 1));
-                                    } else {
-                                        dateFormatterPattern = patternAndTZid;
-                                        timeZone = ZoneOffset.UTC;
-                                    }
-                                    dateFormatter = DateFormatter.forPattern(dateFormatterPattern);
-                                }
-
-                                DateFormatter formatter = dateFormatter.withZone(timeZone);
-                                DateMathParser dateMathParser = formatter.toDateMathParser();
-                                Instant instant = dateMathParser.parse(mathExpression, context::getStartTime, false, timeZone);
-
-                                String time = formatter.format(instant);
-                                beforePlaceHolderSb.append(time);
-                                inPlaceHolderSb = new StringBuilder();
-                                inPlaceHolder = false;
-                            }
-                            break;
-
-                        default:
-                            inPlaceHolderSb.append(c);
-                    }
-                } else {
-                    switch (c) {
-                        case LEFT_BOUND:
-                            if (escapedChar) {
-                                beforePlaceHolderSb.append(c);
-                            } else {
-                                inPlaceHolder = true;
-                            }
-                            break;
-
-                        case RIGHT_BOUND:
-                            if (!escapedChar) {
-                                throw new ElasticsearchParseException("invalid dynamic name expression [{}]." +
-                                    " invalid character at position [{}]. `{` and `}` are reserved characters and" +
-                                    " should be escaped when used as part of the index name using `\\` (e.g. `\\{text\\}`)",
-                                    new String(text, from, length), i);
-                            }
-                        default:
-                            beforePlaceHolderSb.append(c);
-                    }
-                }
-            }
-
-            if (inPlaceHolder) {
-                throw new ElasticsearchParseException("invalid dynamic name expression [{}]. date math placeholder is open ended",
-                    new String(text, from, length));
-            }
-            if (beforePlaceHolderSb.length() == 0) {
-                throw new ElasticsearchParseException("nothing captured");
-            }
-            return beforePlaceHolderSb.toString();
         }
     }
 }
