+++ server/src/test/java/org/elasticsearch/index/query/SearchIndexNameMatcherTests.java
@@ -24,8 +24,8 @@ import org.elasticsearch.cluster.ClusterName;
 import org.elasticsearch.cluster.ClusterState;
 import org.elasticsearch.cluster.metadata.AliasMetaData;
 import org.elasticsearch.cluster.metadata.IndexMetaData;
-import org.elasticsearch.cluster.metadata.IndexNameExpressionResolver;
 import org.elasticsearch.cluster.metadata.MetaData;
+import org.elasticsearch.cluster.metadata.TestIndexNameExpressionResolver;
 import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.test.ESTestCase;
@@ -49,8 +49,8 @@ public class SearchIndexNameMatcherTests extends ESTestCase {
         ClusterService clusterService = mock(ClusterService.class);
         when(clusterService.state()).thenReturn(state);
 
-        matcher = new SearchIndexNameMatcher("index1", "", clusterService, new IndexNameExpressionResolver());
-        remoteMatcher = new SearchIndexNameMatcher("index1", "cluster", clusterService, new IndexNameExpressionResolver());
+        matcher = new SearchIndexNameMatcher("index1", "", clusterService, new TestIndexNameExpressionResolver());
+        remoteMatcher = new SearchIndexNameMatcher("index1", "cluster", clusterService, new TestIndexNameExpressionResolver());
     }
 
     private static IndexMetaData.Builder indexBuilder(String index) {
