+++ x-pack/plugin/enrich/src/test/java/org/elasticsearch/xpack/enrich/EnrichPolicyExecutorTests.java
@@ -6,16 +6,11 @@
 
 package org.elasticsearch.xpack.enrich;
 
-import java.util.Collections;
-import java.util.List;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-import java.util.function.LongSupplier;
-
 import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.LatchedActionListener;
 import org.elasticsearch.client.Client;
 import org.elasticsearch.cluster.metadata.IndexNameExpressionResolver;
+import org.elasticsearch.cluster.metadata.TestIndexNameExpressionResolver;
 import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.util.concurrent.EsRejectedExecutionException;
@@ -28,6 +23,12 @@ import org.elasticsearch.xpack.core.enrich.action.ExecuteEnrichPolicyAction;
 import org.elasticsearch.xpack.core.enrich.action.ExecuteEnrichPolicyStatus;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
+import java.util.function.LongSupplier;
 
 import static org.hamcrest.CoreMatchers.containsString;
 
@@ -141,7 +142,7 @@ public class EnrichPolicyExecutorTests extends ESTestCase {
             null,
             testTaskManager,
             testThreadPool,
-            new IndexNameExpressionResolver(),
+            new TestIndexNameExpressionResolver(),
             ESTestCase::randomNonNegativeLong
         );
 
@@ -198,7 +199,7 @@ public class EnrichPolicyExecutorTests extends ESTestCase {
             null,
             testTaskManager,
             testThreadPool,
-            new IndexNameExpressionResolver(),
+            new TestIndexNameExpressionResolver(),
             ESTestCase::randomNonNegativeLong
         );
